requirements.txt
aniso8601==1.1.0
Flask==0.10.1
Flask-RESTful==0.3.4
Flask-Testing==0.4.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
nose==1.3.7
python-dateutil==2.4.2
pytz==2015.7
regex==2015.11.14
requests==2.8.1
six==1.10.0
SQLAlchemy==1.0.9
Werkzeug==0.11
wheel==0.24.0
